+++ torch/csrc/autograd/functions/convolution.h
@@ -42,7 +42,7 @@ struct ConvForward : public Function, public ConvParams, public HasPrimSpec {
 
   virtual std::string name() override;
   virtual variable_list apply(const variable_list& inputs) override;
-  virtual void primspec(PrimSpecContext* ctx, jit::node_list inputs, jit::node_list outputs);
+  HAS_PRIMSPEC;
 
   std::vector<long> output_size(at::Tensor& input, at::Tensor& weight);
 };
